+++ src/main/java/uk/org/okapibarcode/backend/MsiPlessey.java
@@ -22,7 +22,7 @@ package uk.org.okapibarcode.backend;
 public class MsiPlessey extends Symbol {
 
     private final String MSI_PlessTable[] = {
-        "12121212", "12121221", "12122112", "12122121", "12211212", "12211221", 
+        "12121212", "12121221", "12122112", "12122121", "12211212", "12211221",
         "12212112", "12212121", "21121212", "21121221"
     };
 
@@ -39,7 +39,7 @@ public class MsiPlessey extends Symbol {
         int weight;
         int checkDigit1;
         int checkDigit2;
-        
+
         if (!(content.matches("[0-9]+?"))) {
             error_msg = "Invalid characters in input";
             return false;
@@ -49,9 +49,9 @@ public class MsiPlessey extends Symbol {
         for (i = 0; i < length; i++) {
             intermediate += MSI_PlessTable[Character.getNumericValue(content.charAt(i))];
         }
-        
+
         readable = content;
-        
+
         if ((option2 == 1) || (option2 == 2)) {
             /* Add Modulo-10 check digit */
             evenString = "";
@@ -74,48 +74,48 @@ public class MsiPlessey extends Symbol {
                     }
                 }
             }
-            
+
             if (oddString.length() == 0) {
                 addupString = "0";
             } else {
                 addupString = Integer.toString(Integer.parseInt(oddString) * 2);
             }
-            
+
             addupString += evenString;
-            
+
             addup = 0;
             for(i = 0; i < addupString.length(); i++) {
-                addup += (int)(addupString.charAt(i) - '0');
+                addup += addupString.charAt(i) - '0';
             }
-            
+
             checkDigit1 = 10 - (addup % 10);
             if (checkDigit1 == 10) {
                 checkDigit1 = 0;
             }
-            
+
             intermediate += MSI_PlessTable[checkDigit1];
             readable += checkDigit1;
         }
-        
+
         if ((option2 == 3) || (option2 == 4)) {
             /* Add a Modulo-11 check digit */
             weight = 2;
             addup = 0;
             for (i = content.length() - 1; i >= 0; i--) {
-                addup += (int)(content.charAt(i) - '0') * weight;
+                addup += (content.charAt(i) - '0') * weight;
                 weight++;
-                
+
                 if (weight == 8) {
                     weight = 2;
                 }
             }
-            
+
             checkDigit1 = 11 - (addup % 11);
-            
+
             if (checkDigit1 == 11) {
                 checkDigit1 = 0;
             }
-            
+
             readable += checkDigit1;
             if (checkDigit1 == 10) {
                 intermediate += MSI_PlessTable[1];
@@ -124,7 +124,7 @@ public class MsiPlessey extends Symbol {
                 intermediate += MSI_PlessTable[checkDigit1];
             }
         }
-        
+
         if ((option2 == 2) || (option2 == 4)) {
             /* Add a second Modulo-10 check digit */
             evenString = "";
@@ -147,31 +147,31 @@ public class MsiPlessey extends Symbol {
                     }
                 }
             }
-            
+
             if(oddString.length() == 0) {
                 addupString = "0";
             } else {
                 addupString = Integer.toString(Integer.parseInt(oddString) * 2);
             }
-            
+
             addupString += evenString;
-            
+
             addup = 0;
             for(i = 0; i < addupString.length(); i++) {
-                addup += (int)(addupString.charAt(i) - '0');
+                addup += addupString.charAt(i) - '0';
             }
-            
+
             checkDigit2 = 10 - (addup % 10);
             if (checkDigit2 == 10) {
                 checkDigit2 = 0;
             }
-            
+
             intermediate += MSI_PlessTable[checkDigit2];
             readable += checkDigit2;
         }
-        
+
         intermediate += "121"; // Stop
-        
+
         pattern = new String[1];
         pattern[0] = intermediate;
         row_count = 1;
